% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_network_tools.R
\name{adjust_intra_sims}
\alias{adjust_intra_sims}
\title{Adjust intra-group similarities based on inter-group ones}
\usage{
adjust_intra_sims(PSN, groups)
}
\arguments{
\item{PSN}{Patient Similarity Network in Adjacent Matrix format}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}
}
\value{
Return the patient similarity with the weights between patients adjusted
}
\description{
Considers the similarity between two patients belonging to the same group, it increases this similarity if
they have low similarities with members of other groups, it lowers down this similarity if the opposite
}
