% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_wrappers.R
\name{analyse_training}
\alias{analyse_training}
\title{Analyse training patients to find significant PSNs}
\usage{
analyse_training(
  data_l,
  max_size = 200,
  min_elements = 5,
  n_top_pathways = 100,
  n_cores = NULL,
  keep_PSN = FALSE,
  ban = TRUE
)
}
\arguments{
\item{data_l}{The output list obtained from processing of the user's data with the function: prepare_data}

\item{max_size}{Default 150, integer value filtering out the pathways with more elements than this threshold}

\item{min_elements}{Default 4, integer value filtering out the pathways containing less SIGNIFICANT molecules than this threshold}

\item{n_cores}{Default 2, integer value that defines the number of cores to use for running the function in parallel}

\item{keep_PSN}{Default FALSE, boolean that specifies to not keep the PSN generated from significantly enriched pathways}

\item{ban}{Default TRUE, remove excessive context-specific pathways (e.g. "SOUND", "STABILIZATION", "ODONTOGENESIS", "PREGNANCY")}

\item{n_top_sets}{Default 50, integer value that defines how many best pathways-specific PSNs to return as result}
}
\value{
For each unique omic (e.g. gene expression), it provides:

- the normalized count matrix

- the SDR dataframe (obtained with find_SDR()) including the statistics of the molecules tested for separating the patient's groups.

It also provide the result of the pathway analysis that combines the information of the different omics.
Precisely, a list of two elements accessible with $training$pathway_analysis: top_pathways_l and top_pathways_df.

- top_pathways_df contains meta-information about the significantly enriched pathways which have been used to build a significant PSN.

- top_pathways_l is the list s.t. each element refers to a significantly enriched pathway and contains:

1) the vector of significant molecules enriching the pathway

2) the count matrix used to build the PSN

3) statistics related to the PSN used to model the relationships between patients based on the specific pathway info
}
\description{
For each omic, it determines the best molecules to build the PSNs.
The molecules are fit into the pathways or regulatory target sets.
StellarPath finds the most enriched pathways.
A PSN is built for each pathway and is tested if separates the groups in comparison.
If yes, the PSN is kept and used for the training and classification.
}
