% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_similarities.R
\name{build_avg_PWN}
\alias{build_avg_PWN}
\title{Create Similarity Network based on Average of Pair Weights}
\usage{
build_avg_PWN(m, comb2 = NULL)
}
\arguments{
\item{m}{Numeric matrix s.t. columns are objects, rows are features and a cell contains a value}
}
\value{
Similarity Network based on Average of Pair Weights
}
\description{
For each pair of column profiles, it determines the average of the feature's values and then
it determines their average to assign the score of similarity to the pair of profiles.
}
