% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_similarities.R
\name{build_PSN}
\alias{build_PSN}
\title{Create Patient Similarity Network with StellarPath's similarity}
\usage{
build_PSN(m, up = TRUE)
}
\arguments{
\item{m}{Numeric matrix s.t. columns are objects, rows are features and a cell contains a value}

\item{up}{Default TRUE, boolean that defines if the significant molecules are upregulated in a patient's group.}
}
\value{
Similarity Network based on Adjusted Weighted Jaccard similarity and Average of Pair Weights
}
\description{
It creates the PSN based on the Adjusted Weighted Jaccard similarity.
It creates the PSN based on the Average of Pair Weights.
In case of upregulated significant molecules populating the count matrix, then more
the average is high and more the similarity is high.
In case of downregulated significant molecules populating the count matrix, then more
the average is low and more the similarity is high.
It combines the two values that each pair of samples has.
The combination is performed computing the triangular area of the two values of similarity.
}
