% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_similarities.R
\name{build_PSNs}
\alias{build_PSNs}
\title{Create a Similarity Network based on Adjusted Weighted Jaccard similarity and Average of Pair Weights}
\usage{
build_PSNs(m)
}
\arguments{
\item{m}{Numeric matrix s.t. columns are subjects, rows are features and a cell contains a value}
}
\value{
Similarity Network based on Adjusted Weighted Jaccard similarity and Average of Pair Weights
}
\description{
It creates the SN based on the Adjusted Weighted Jaccard similarity
It creates the SN based on the Average of Pair Weights
It combines the two values that each pair of samples has
The combination is performed computing the triangular area of the two values
}
