% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{check_vectors}
\alias{check_vectors}
\title{Checks vectors of elements}
\usage{
check_vectors(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
The result of the check. 0 not match. 1 match.
}
\description{
This function checks if two vectors are exactly the same both in the elements that contain and their order
}
