% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\docType{data}
\name{cll_l}
\alias{cll_l}
\title{List of data describing cll_l samples vs non-treated ones}
\format{
The list has three elements
\describe{
  \item{info}{dataframe containing information about the mouse samples}
  \item{mRNA}{matrix containing the bulk RNA counts (both genes and lncRNAs)}
  \item{miRNA}{matrix containing the miRNA counts}
}
}
\source{
{
  \url{}
}
}
\usage{
cll_l
}
\description{
List of data describing cll_l samples vs non-treated ones
}
\keyword{datasets}
