% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{combine_test_results}
\alias{combine_test_results}
\title{Combine results from testing pathways with ECDF and PSN}
\usage{
combine_test_results(el1, el2, target_name = "test")
}
\arguments{
\item{el1}{edge list containing elements connections that have been found significant based on PSN}

\item{el2}{edge list containing elements connections that have been found significant based on ECDF}
}
\value{
edge list containing elements connections that have been found significant based on PSN/ECDF and PSN+ECDF
}
\description{
Combine results from testing pathways with ECDF and PSN
}
