% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_network_tools.R
\name{convert_edg2adj}
\alias{convert_edg2adj}
\title{Converts edge list to adjacency matrix}
\usage{
convert_edg2adj(el)
}
\arguments{
\item{el}{edge list of a graph}
}
\value{
Return djacency matrix of the original network as matrix
}
\description{
Converts a graph represented as edge list to adjacency matrix
}
