% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_monitor.R
\name{convert_min2h}
\alias{convert_min2h}
\title{Convert minutes time value in hours}
\usage{
convert_min2h(raw_min)
}
\arguments{
\item{raw_min}{Double value}
}
\value{
New double value formatted in hours
}
\description{
Convert minutes time value in hours
}
