% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{create_df4attr}
\alias{create_df4attr}
\title{Create a dataframe from a vector of strings containing attribute-value pairs}
\usage{
create_df4attr(strings)
}
\arguments{
\item{strings}{A vector of strings containing attribute-value pairs}
}
\value{
A dataframe with attribute names as columns and attribute values assigned to the respective rows
}
\description{
This function takes a vector of strings, where each string represents multiple attribute-value pairs separated by '|'. The attribute-value pairs are in the format 'attribute;;value'. The function extracts the attribute names and values and creates a dataframe where each attribute becomes a column and the values are assigned to the corresponding rows.
}
\examples{
strings <- c("age at death;;93|braak score;;IV|diagnosis;;Alzheimer's disease",
             "age at death;;85|braak score;;II|diagnosis;;Parkinson's disease",
             "age at death;;78|braak score;;III|diagnosis;;Alzheimer's disease")
df <- create_df4attr(strings)
print(df)
}
