% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\docType{data}
\name{dbs_mmu_net_l}
\alias{dbs_mmu_net_l}
\title{Mouse gene-gene interaction network}
\format{
list
\describe{
  \item{NETWORK}{gene-gene interaction network in adjacent and edge list format}
}
}
\source{
{
  \url{https://www.ebi.ac.uk/intact/home}
}
}
\usage{
dbs_mmu_net_l
}
\description{
Mouse gene-gene interaction network
}
\keyword{datasets}
