% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\docType{data}
\name{dbs_mmu_paths_l}
\alias{dbs_mmu_paths_l}
\title{Human specific list of pathways about different omics}
\format{
The list has two levels, first level indicates the specie, second level indicates the set of database data:
\describe{
  \item{MIRWALK}{list where for each miRNA contains the set of its target genes}
  \item{RNARNA}{list that contains RNA-RNA interactions (e.g. lncrna_gene_targ_l contains the target genes for each lncRNA)}
  \item{PATHWAYS}{list that contains gene canonical pathways}
}
}
\source{
{
  \url{http://mirwalk.umm.uni-heidelberg.de/}
  \url{http://www.rnainter.org/}
  \url{http://baderlab.org/GeneSets}
}
}
\usage{
dbs_mmu_paths_l
}
\description{
Human specific list of pathways about different omics
}
\keyword{datasets}
