% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{do_complete_net}
\alias{do_complete_net}
\title{Complete the network of multi-omics molecules}
\usage{
do_complete_net(df, pathway_l)
}
\arguments{
\item{df}{character dataframe, first column contains the pathway names, second column contains the molecules enriching the corresponding row pathway}

\item{pathway_l}{list of pathways with the names that match the ones present in the first column of the dataframe}
}
\value{
Updated dataframe
}
\description{
A mutation can alter a pathway.
A lncRNA can alter another pathway.
The two pathways can overlap.
This function introduces new connections to the edge list dataframe.
A row to connect the mutation to the lncRNA's pathway.
A row for the opposite.
}
