% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{do_standard}
\alias{do_standard}
\title{0-1 standardization}
\usage{
do_standard(m, min = 0, max = 1, capping = FALSE)
}
\arguments{
\item{m}{numeric matrix}

\item{min}{Default 0, number to set the low limit of the range standardization}

\item{max}{Default 1,  number to set the high limit of the range standardization}

\item{capping}{Default FALSE, do not perform capping before standardize}
}
\value{
range standardized input matrix
}
\description{
Vectorize the matrix, standardize based on 0 and 1, return the standardized matrix
}
