% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_counts.R
\name{do_winz}
\alias{do_winz}
\title{Winsorizing}
\usage{
do_winz(m, groups, min = 0.05, max = 0.95)
}
\arguments{
\item{m}{numeric matrix of samples divided in groups at the columns}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}
}
\value{
Input winzorised numeric matrix
}
\description{
This function performs Winsorizing of the row values based on column groups
}
