% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{find_cdf_pathways}
\alias{find_cdf_pathways}
\title{Find significant regulatory targets with ecdf test}
\usage{
find_cdf_pathways(
  nc_l,
  nc_sign_l,
  nc_tT_ov,
  target_tT_ov,
  pathway_type = "generic_pathway",
  count_source = "sequencing",
  anti_corr = FALSE
)
}
\arguments{
\item{nc_l}{list of sets, each set called as regulatory element contains the names of the element's targets}

\item{nc_sign_l}{list of significant sets, each set called as deregulated regulatory element contains the names of the element's deregulated targets}

\item{nc_tT_ov}{dataframe of statics regarding the regulatory elements, elements must be at the rows and there must be a column logFoldChange}

\item{target_tT_ov}{dataframe of statics regarding the regulatory elements' targets, targets must be at the rows and there must be a column logFoldChange}

\item{pathway_type}{Default generic_pathway, label indicating the type of pathways (e.g. "gene_canonical_pathways")}

\item{count_source}{label sequencing, label indicating the source used to find the significant molecules (e.g. "gene_expression", "mutation")}

\item{anti_corr}{default FALSE, indicates if the regulatory element and its targets must be anti-correlated in fold changes}
}
\value{
dataframe with two columns: regulatory element (e.g. miRNA) and its significant target/s
}
\description{
Find the log fold change of the targets and non-targets.
Test if the two lists of fold changes are statistically different using ks.test.
Determine if the targets are enriching positively or negatively with respect the non-targets.
For each set of statistically significant targets, return the regulatory element (e.g. miRNA) with its targets.
}
