% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{find_mutated_pathways}
\alias{find_mutated_pathways}
\title{Find all enriched pathways (dangerous)}
\usage{
find_mutated_pathways(
  pathways_l,
  sign_df,
  gex,
  pathway_type = "generic_pathway",
  count_source = "sequencing",
  max_size = 200,
  min_elements = 1,
  n_top_pathways = 200,
  n_cores = 2,
  keep_PSN = FALSE,
  ban = TRUE
)
}
\arguments{
\item{pathways_l}{list of pathways}

\item{sign_df}{dataframe generated by find_SDR}

\item{pathway_type}{Default generic_pathway, label indicating the type of pathways (e.g. "gene_canonical_pathways")}

\item{count_source}{label sequencing, label indicating the source used to find the significant molecules (e.g. "gene_expression", "mutation")}

\item{max_size}{Default 200, integer filtering out the pathways with more elements than this threshold}

\item{min_elements}{Default 1, integer filtering out the pathways containing less than this threshold of significant molecules}

\item{n_top_pathways}{Default 100, integer setting how many best pathway-specific PSNs to retrieve}

\item{n_cores}{Default 2, integer setting the number of cores to use for running the function in parallel}

\item{keep_PSN}{Default FALSE, boolean, do not keep the PSN generated from a pathway}

\item{ban}{Default TRUE, boolean, remove excessive context-specific pathways (e.g. "SOUND", "STABILIZATION", "ODONTOGENESIS", "PREGNANCY")}
}
\value{
list of two elements: top_pathways_l and top_pathways_df.
top_pathways_df details the pathways enriched by the significant molecules.
It describes them with the following column features:

- pathways: pathway name

- molecules: significantly different molecule between groups that has been found associated to the enriched pathway

- barcode: direction of deregulation of the molecule

While, top_pathways_l contains the enriched pathways with all their annotated molecules
}
\description{
Remove the excessive context-specific pathways (ban_list).
Remove the pathways which size is bigger than max_size (max_size:200).
Keep all the pathways having significant elements (pathway_df).
}
