% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{find_PSN}
\alias{find_PSN}
\title{Find which pathway/set of elements are represented by a significant similarity network}
\usage{
find_PSN(
  counts,
  groups,
  pathways_l,
  sign_df,
  max_size = 200,
  min_elements = 6,
  n_top_pathways = 100,
  n_cores = 2,
  keep_PSN = FALSE
)
}
\arguments{
\item{counts}{count matrix of samples at the columns and molecules at the row (e.g. gene expression matrix)}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}

\item{pathways_l}{list of pathways or generically set of elements}

\item{sign_df}{dataframe s.t. rownames are molecules's names belonging to the list of pathways, column logFoldChange
with double values indicating log fold change of the molecule's count value between two groups, column Significant
which is TRUE or FALSE based on if the row molecule has been tested and resulted signficant, column Barcode has a
0 value for not significant molecules and a 9 for significant values.}

\item{max_size}{Default 150, integer value filtering out the pathways/sets with more elements than this threshold}

\item{min_elements}{Default 4, integer value filtering out the pathways/sets containing less than this threshold of significant molecules}

\item{n_cores}{Default 2, integer that set the number of cores to use for running the function in parallel}

\item{keep_PSN}{Default FALSE, do not keep the PSN generated from a set of element}

\item{n_top_sets}{Default 50, integer that set how many best sets/PSNs to retrieve}
}
\value{
list of two elements: top_pathways_l and top_pathways_df.
top_pathways_df details the pathways/sets which have a significant PSN.
It describes them with the following column features:
- pathway name
- freq refers to the number of significant molecules/elements belonging to the pathway
- regulation refers to the direction of enrichment, in case the molecules belonging to the pathway have count values
which are greater in the strongest group than in the opposite group of the network then it is "activated"
- groups indicates the group of nodes being significantly more similar than the opposite group in comparison
- power indicates from 1 to 10 how much the strongest group is more similar than the non-members
- min_intra_STRONG is the minimum similarity between the members of the strongest group which is greater than the opposite group's intrasimilarities and the intergroup similarities
- max_intra_WEAK is the maximum similarity between the members of the weakest group which is lower than the opposite group's intrasimilarities and the intergroup similarities
- max_inter_similarities is the maximum similarity between members of the two groups
While, top_pathways_l contains the data associated to each pathway/set of element that has been tested.
- set contains the significant molecules/elements belonging to the pathway/set
- set_counts is the count matrix (e.g. genes per samples) with only the count values of the significant molecules/elements belonging to the pathway/set
- stats is a dataframe indicating the statistics of the pathway/set
- centrality is a dataframe s.t. for each sample indicates how much is strongly similar in its groups vs its not group
- orig_set contains all the original elements of the pathway/set
}
\description{
Remove the pathways/sets which size is bigger than max_size.
For each set of elements (pathways_l), keep only the significant elements (sign_df).
Keep the (n_top_pathways) having the highest number of significant elements (sign_df).
For each set of significant elements (pathways_l), use their values (counts) to compute how much each pair
of samples is similar and build the similarity network.
Compute the Power of separation for each network.
}
