% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_counts.R
\name{find_SDR_directional}
\alias{find_SDR_directional}
\title{Find directional specific differential role/expressed molecules between two groups}
\usage{
find_SDR_directional(counts, groups, top_molecules_th = 0.05)
}
\arguments{
\item{counts}{count matrix of samples at the columns and molecules at the row (e.g. gene expression matrix)}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}

\item{top_molecules_th}{Default 0.1, double value of the molecules that are selected as best ones}
}
\value{
Dataframe of descriptive statistics related to the row molecules in a count matrix having two groups of samples in comparison
logFoldChange measures the change in count values between first vs second group defined in the groups variable (unique(groups)[1] vs unique(groups)[2])
Stability measures how much is differently stable in first vs second group (higher positive and more is stable for first group)
Activity_1 shows the mean of the count values (rank or expression) in the first group
Activity_2 shows the mean of the count values (rank or expression) in the second group
Stable_for indicates for which sample group the molecule is stable for
Significant indicates if the molecule has been tested significant
Barcode is an internal variable used by the software to understand which test the molecule has passed to be significant
}
\description{
Internal function of find_SDR
}
