% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_initialize.R
\name{find_SP}
\alias{find_SP}
\title{Find StellarPath required python packages}
\usage{
find_SP(ask)
}
\arguments{
\item{ask}{logical; if \code{FALSE}, use the first StellarPath required python packages installation found;
if \code{TRUE}, list available StellarPath required python packages installations and prompt the user
for which to use. If another (e.g. \code{python_executable}) is set, then
this value will always be treated as \code{FALSE}.}
}
\value{
SP_python
}
\description{
Locate the user's version of Python for which StellarPath required python packages are installed.
}
\keyword{internal}
