% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_initialize.R
\name{find_SP_env}
\alias{find_SP_env}
\title{Find StellarPath required python pacakges env}
\usage{
find_SP_env()
}
\description{
check whether conda/virtual environment for StellarPath required python pacakges exists
}
\keyword{internal}
