% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{get_auc}
\alias{get_auc}
\title{Calculates the Area Under the Receiver Operating Characteristic (ROC) Curve (AUC).}
\usage{
get_auc(true_labels, predicted_scores)
}
\arguments{
\item{true_labels}{A vector of true binary labels}

\item{predicted_scores}{A vector of predicted scores or probabilities for the positive class.}
}
\value{
A numeric value representing the AUC.
}
\description{
This function takes in a vector of true labels and a vector of predicted scores/probabilities, and returns the AUC using the roc() function from the pROC package.
}
\examples{
true_labels=c("LATE","LATE","LATE","EARLY","EARLY","EARLY")
predicted_scores=c(0.8,0.9,0.2,0.4,0.3,0.9)
get_auc(true_labels, predicted_scores)
}
