% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_similarities.R
\name{get_AUquad}
\alias{get_AUquad}
\title{Quadrilateral area between four points}
\usage{
get_AUquad(b_value, c_value, d_value)
}
\arguments{
\item{b_value}{Value}

\item{c_value}{Value}

\item{d_value}{Value}
}
\value{
Triangle area
}
\description{
One is fixed (0,0), the second is (b_value,0), the third is (0,c_value), the fourth is (b_value,c_value).
Quadrilateral area increases more when three points increase.
Quadrilateral area penalizes when only one point changes.
}
