% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_network_tools.R
\name{get_cohesive_score}
\alias{get_cohesive_score}
\title{Get a cohesive centrality score for each patient in PSN}
\usage{
get_cohesive_score(PSN, groups)
}
\arguments{
\item{PSN}{Patient Similarity Network in Adjacent Matrix format}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}
}
\value{
Matrix of values, first column indicates the median of the intraclass similarities,
}
\description{
Function which computes how much a patient/node is connected to the members of its group, its non-members and then
it produces a centrality score. Higher is the score and more a node is strongly connected to the members of its
group and lowly connected to the non-members.
}
