% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_counts.R
\name{get_lcpm}
\alias{get_lcpm}
\title{Library normalize a raw count matrix}
\usage{
get_lcpm(
  counts,
  groups = NULL,
  filtering = TRUE,
  winz = FALSE,
  min = 0.1,
  max = 0.9
)
}
\arguments{
\item{counts}{Numeric matrix of raw count values (e.g. molecules x samples)}

\item{groups}{Character vector, each element indicates the group of the sample in the same column position in counts}

\item{filtering}{Default TRUE, removing row molecules with low counts on the overall samples}

\item{winz}{Default FALSE, not adjusting outliers by Winsorizing}

\item{min}{Default 0.1, double value indicating the minimum value of the distribution for the Winsorizing,
it must be smaller than 1 and max value}

\item{max}{Default 0.9, double value indicating the maximum value of the distribution for the Winsorizing,
it must be smaller than 1 and greater than min value}
}
\value{
library normalized count matrix in lcpm
}
\description{
This function filters, library normalizes and log transform count values with edgeR
}
