% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_monitor.R
\name{get_max_ram_used}
\alias{get_max_ram_used}
\title{End monitor ram usage in linux and get maximum value}
\usage{
get_max_ram_used(log_path)
}
\arguments{
\item{log_name}{Character string that has been given to the log file saved on the disk with start_monitor_ram}
}
\value{
Maximum ram used and collected in the log file
}
\description{
End monitor ram usage in linux and get maximum value
}
