% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{get_mcc}
\alias{get_mcc}
\title{Matthews correlation coefficient (MCC) score}
\usage{
get_mcc(real, predicted, one_group)
}
\arguments{
\item{act}{actual values (vector), 1 (positive), or 0 (negative)}

\item{pred}{predict values (vector), 1 (positive), or 0 (negative)}
}
\value{
MCC score
}
\description{
Calculate the Matthews correlation coefficient (MCC) score
}
