% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{get_meanXgr}
\alias{get_meanXgr}
\title{Compute mean per group}
\usage{
get_meanXgr(m, groups, median = FALSE)
}
\arguments{
\item{m}{numeric matrix}

\item{groups}{character or factor vector s.t. one element is the label indicating the group of one sample column}

\item{median}{boolean, default is FALSE, set to true if you want to use the median instead of the mean}
}
\value{
numeric matrix wit same column of m, number of columns equal to the number of groups, a cell's value is
equal to the mean of the values of the row feature in the original columns of the same group
}
\description{
Compute mean per column group of each feature row
}
