% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{get_pathway_data}
\alias{get_pathway_data}
\title{Retrieve pathway's data from path}
\usage{
get_pathway_data(data_l, path2path)
}
\arguments{
\item{data_l}{omics list}

\item{path2path}{character string pointing to the position of the omics list in the data list}
}
\value{
list of two elements: top_pathways_df and top_pathways_el with pathways' information produced with different omics
}
\description{
Retrieve pathway's data from path
}
