% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{get_perf_metrics}
\alias{get_perf_metrics}
\title{Calculates the classification metrics given the true and predicted labels.}
\usage{
get_perf_metrics(true_labels, predicted_labels)
}
\arguments{
\item{true_labels}{A vector of true binary labels}

\item{predicted_labels}{A vector of predicted binary labels}
}
\value{
A dataframe containing the calculated metrics: Matthews correlation coefficient, accuracy, sensitivity, and specificity.
}
\description{
This function takes in a vector of true binary labels and a vector of predicted binary labels.
Returns the classification metrics (Matthews correlation coefficient, accuracy, sensitivity, and specificity).
}
\examples{
true_labels=c("LATE","LATE","LATE","EARLY","EARLY","EARLY")
predicted_labels=c("LATE","LATE","EARLY","EARLY","EARLY","LATE")
get_perf_metrics(true_labels, predicted_labels)
}
