% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_network_tools.R
\name{get_power}
\alias{get_power}
\title{Determines the Power of a Patient Similarity Network}
\usage{
get_power(PSN, groups)
}
\arguments{
\item{PSN}{Patient Similarity Network in Adjacent matrix format}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}
}
\value{
Dataframe indicating the group which nodes are more similar than the opposite, the power or magnitude measuring
how much the strongest group is more similar than the others, the smallest similarity of the strongest group which is greater
than the biggest similarity of the opposite group and of the similarities between the groups. Dataframe is NA if
the power is 0 and the two groups do not show significant different trends of similarities.
}
\description{
Determines if the two groups of nodes in the network separate and how much
}
\details{
More a group has the intragroup similarities greater than the opposite intragroup similarities and
intergroup similarities and more the power is close to 10

In order to understand the magnitude, a low percentile (e.g. 0.2) of the similarities of a group
is tested if is greater than a high percentile (e.g. 0.8) of the other two sets of similarities.

Lower is the low percentile and higher is the high percentile while still keeping a separability of
the similarities and more the power is high.
}
