% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{get_sdXgr}
\alias{get_sdXgr}
\title{Compute standard deviation per group}
\usage{
get_sdXgr(m, groups, mad = FALSE)
}
\arguments{
\item{m}{numeric matrix}

\item{groups}{character or factor vector s.t. one element is the label indicating the group of one sample column}

\item{mad}{boolean, default is FALSE, set to true if you want to use median absolute deviations}
}
\value{
numeric matrix wit same column of m, number of columns equal to the number of groups, a cell's value is
equal to the standard deviation of the values of the row feature in the original columns of the same group
}
\description{
Compute standard deviation per column group of each feature row
}
