% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_network_tools.R
\name{get_subgraph}
\alias{get_subgraph}
\title{Extract subgraph from edge list}
\usage{
get_subgraph(el, vx, order_neigh = 1)
}
\arguments{
\item{el}{edge list representing the graph, with the first two columns indicating the name of the connected nodes, and any additional columns containing meta-information related to each edge.}

\item{vx}{name of the node of interest in the graph}

\item{order_neigh}{Default 1, integer indicating the order of the neighbors to include in the subgraph from the node of interest}
}
\value{
A data frame containing the edges of the subgraph formed by the node of interest and its neighbors, along with the information from the extra columns.
}
\description{
This function takes an edge list representing a graph and the name of a node of interest and returns the subgraph formed by the node of interest and its neighbors, along with the information from the extra columns.
}
