% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_propagation.R
\name{harm_countsWnet}
\alias{harm_countsWnet}
\title{Harmonizes the matrix of sample's profiles with respect the network}
\usage{
harm_countsWnet(counts, net)
}
\arguments{
\item{counts}{Numeric matrix with rows matching nodes of the network}

\item{net}{Adjacency matrix of a network s.t. nodes have same names of the rows in the profile matrix}
}
\value{
The matrix of sample's profiles without the rows not having a match with a node of the network
}
\description{
Remove all the features/rows that don't have their name matching with a node inside the network
}
