% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_wrappers.R
\name{infer_new_data}
\alias{infer_new_data}
\title{Infer the class of new patients and find significant PSNs}
\usage{
infer_new_data(
  info,
  groups_name = NULL,
  gex = NULL,
  lncRNA = NULL,
  miRNA = NULL,
  mutations = NULL,
  tax_id = 9606,
  simplify_names = TRUE,
  cv_probs = c(0.7, 0.1, 0.2),
  user_sets = NULL,
  data_l = NULL,
  n_cores = 2,
  seed = 5,
  winz = FALSE,
  type = "sample_rank"
)
}
\arguments{
\item{info}{Dataframe, two character columns, first has sample's IDs (have to match with column names of the count matrices),
second has sample's groups (only two as for a pairwise DE analysis, for example c(AD,AD,AD,HEALTHY,HEALTHY,HEALTHY))}

\item{groups_name}{Character vector of length two, it must contain the same labels of unique(info[,2]), tip: first the case group like AD,
while the second name refers to the control group like HEALTHY}

\item{gex}{Numeric matrix of raw count values from bulk RNA sequencing (e.g. genes x samples) (it may include also lncRNAs)}

\item{lncRNA}{Numeric matrix of raw count values from bulk RNA sequencing (e.g. lncRNA x samples)}

\item{miRNA}{Numeric matrix of raw count values from miRNA sequencing (e.g. miRNA x samples)}

\item{mutations}{Numeric matrix of binary values from somatic mutation data (e.g. gene X samples, 1 per mutated gene, 0 otherwise)}

\item{tax_id}{Default 9606, integer value indicating taxonomic id of the samples, either 9606 (human) or 10090 (mouse)}

\item{simplify_names}{Default TRUE, simplify names of the samples in the info and count matrices}

\item{cv_probs}{Default c(0.7,0.1,0.2), vector of doubles indicating the proportion of the samples to use as training, validation and testing set}

\item{user_sets}{Default NULL, list of elements (one element is used to compose the training and testing set of a cv run) s.t. each element is a list containing two numeric vectors, one called train and one called test, each vector indicate the column index of those patients falling into training or testing set}

\item{data_l}{Default NULL, results list of data produced with StellarPath workflow on a previous patient dataset}

\item{n_cores}{Default 2, integer value greater than 0 indicating the number of cores to use to parallelize and speed up the operations}

\item{seed}{Default 5, integer value indicating the seed to replicate the sampling of the cross validation}

\item{winz}{Default FALSE, boolean to indicate if to adjuste outliers with winsorizing}

\item{type}{Default total rank, character string which defines the type of transformation applied to the count data, options in the help of the get_rank_01 function}
}
\value{
A list s.t. each element contains the user-input prepared data for a run of cross-validation.
The run's data are divided into training and testing sets.
Each set contains a specific subset of sample's information (i.e. info) and normalized count matrices (i.e. omics).
}
\description{
Infer the class of new patients and find significant PSNs
}
