% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{merge_multi_omics_analysis}
\alias{merge_multi_omics_analysis}
\title{Combines multiple results of analysis of training omics}
\usage{
merge_multi_omics_analysis(l2merge)
}
\arguments{
\item{l2merge}{list to merge s.t. each element of the list includes one top_pathways_l and one top_pathways_df}
}
\value{
list of two elements: top_pathways_l and top_pathways_df
}
\description{
Combines multiple results of analysis of training omics
}
