% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{merge_pathway_analysis}
\alias{merge_pathway_analysis}
\title{Combines resulting pathways from multiple integrations}
\usage{
merge_pathway_analysis(omics, start)
}
\arguments{
\item{omics}{omics list}

\item{start}{character string pointing to the position of the omics list in the data list}
}
\value{
list of two elements: top_pathways_df and top_pathways_el with pathways' information produced with different omics
}
\description{
Combines resulting pathways from multiple integrations
Returns a list with one matrix and one edge list dataframe with the pathways' information produced with different omics
}
