% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\docType{data}
\name{ogd_l}
\alias{ogd_l}
\title{OGD WT vs Normoxia WT}
\format{
The list has three elements
\describe{
  \item{info}{dataframe containing information about the mouse samples}
  \item{mRNA}{dataframe containing the mRNA count matrix}
  \item{miRNA}{dataframe containing the miRNA count matrixs}
}
}
\source{
{
  \url{}
}
}
\usage{
ogd_l
}
\description{
OGD WT vs Normoxia WT
}
\keyword{datasets}
