% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plot.R
\name{plot_PSN}
\alias{plot_PSN}
\title{Plot a Patient Similarity Network}
\usage{
plot_PSN(
  info,
  PSN,
  k_clusters = 10,
  edge_threshold = 0.3,
  edge_size = 5,
  label.cex = 3,
  legend.cex = 0.5,
  edge_cols = c("#03c800", "#96b5a3", "#9ab8ff"),
  node_cols = c("#9affb3", "#9aebff"),
  file_path = "PSN",
  file_type = "pdf",
  file_width = 14,
  file_height = 16,
  GLratio = 2
)
}
\arguments{
\item{info}{Dataframe, two character columns, first has sample's IDs (have to match with column names of the count matrices),
second has sample's groups (only two as a pairwise DE analysis, for example c(AD,AD,AD,HEALTHY,HEALTHY,HEALTHY))}

\item{PSN}{Patient Similarity Network in Adjacent matrix format}

\item{k_clusters}{Default 10, number of nodes to visualize for each group}

\item{edge_threshold}{Default 0.3, similarities lower than 0.3 will be hide in order to simplify the plot}

\item{edge_size}{Default 5, size of the edges}

\item{label.cex}{Default 3, size of the node's labels}

\item{legend.cex}{Default 0.5, size of the legend text}

\item{edge_cols}{Defulat c("#03c800","#96b5a3","#9ab8ff"), vector of three exadecimal colours, first colour is used for first group's
intrasimilarities, second colour for intersimilarities and third colour for second group's intrasimilarities}

\item{node_cols}{Default c("#9affb3","#9aebff"), vector of two exadecimal colours, first colour is used for first group's nodes}

\item{file_path}{Default "PSN", indicates the path and name of the plot file}

\item{file_type}{Default "pdf", indicates the format of the plot file}

\item{file_width}{Default 14, integer that indicates the width of the plot file}

\item{file_height}{Default 16, integer that indicates the height of the plot file}

\item{GLratio}{Relative size of the graph compared to the layout. Defaults to 2.5}
}
\value{
Save the plot on the disk and return a list with the qgraph object and the list of the patient's clusters
}
\description{
Plot a Patient Similarity Network
}
