% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_wrappers.R
\name{prepare_data}
\alias{prepare_data}
\title{Check and Prepare the input data}
\usage{
prepare_data(
  info,
  groups_name = NULL,
  mRNA = NULL,
  miRNA = NULL,
  mutations = NULL,
  tax_id = 9606,
  simplify_names = TRUE,
  cv_probs = c(0.7, 0.1, 0.2),
  n_iter_cv = 2,
  n_cores = 2,
  seed = 5,
  winz = FALSE,
  min = 0.1,
  max = 0.9,
  type = "total_rank"
)
}
\arguments{
\item{info}{Dataframe, two character columns, first has sample's IDs (have to match with column names of the count matrices),
second has sample's groups (only two as a pairwise DE analysis, for example c(AD,AD,AD,HEALTHY,HEALTHY,HEALTHY))}

\item{groups_name}{Character vector of length two, it must contain the same labels of unique(info[,2]), tip: first the case group like AD,
while the second name refers to the control group like HEALTHY}

\item{mRNA}{Numeric matrix of raw count values from mRNA sequencing (e.g. genes x samples)}

\item{miRNA}{Numeric matrix of raw count values from miRNA sequencing (e.g. miRNA x samples)}

\item{mutations}{Numeric matrix of binary values from somatic mutation data (e.g. gene X samples, 1 per mutated gene, 0 otherwise)}

\item{tax_id}{Default 9606, integer value indicating taxonomic id of the samples, either 9606 (human) or 10090 (mouse)}

\item{simplify_names}{Default TRUE, simplify names of the samples in the info and count matrices}

\item{cv_probs}{Default c(0.8,0.1,0.1), vector of doubles indicating the proportion of the samples to use as training, validation and testing set}

\item{n_iter_cv}{Default 2, integer value indicating the number of iterations to perform the cross validation}

\item{n_cores}{Default 2, integer value greater than 0 indicating the number of cores to use to parallelize and speed up the software's operations}

\item{seed}{Default 5, integer value indicating the seed to replicate the sampling of the cross validation}

\item{winz}{Default FALSE, not adjusting outliers by Winsorizing}

\item{min}{Default 0.1, double value indicating the minimum value of the distribution for the Winsorizing,
it must be smaller than 1 and max value}

\item{max}{Default 0.9, double value indicating the maximum value of the distribution for the Winsorizing,
it must be smaller than 1 and greater than min value}

\item{type}{Default expression, character string which defines the type of transformation applied to the count data, options in the help of the get_rank_01 function}
}
\value{
A list s.t. each element contains the user-input prepared data for a run of cross-validation.
The run's data are divided into training and testing sets.
Each set contains a specific subset of sample's information (info) and count values (omics).
}
\description{
This function checks that the input data satisfy all the criteria for being used in the software
It creates an object list that includes all the checked and prepared user data
}
