% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_counts.R
\name{prepare_RNAseq}
\alias{prepare_RNAseq}
\title{Normalization and Rank transformation of RNAseq data}
\usage{
prepare_RNAseq(
  counts,
  groups = NULL,
  winz = FALSE,
  min = 0.1,
  max = 0.9,
  type = "total_rank"
)
}
\arguments{
\item{counts}{Numeric matrix of raw count values from RNA sequencing (e.g. genes x samples)}

\item{groups}{Character vector, each element indicates the group of the sample in the same column position in m}

\item{winz}{Default FALSE, not adjusting outliers by Winsorizing}

\item{min}{Default 0.1, double value indicating the minimum value of the distribution for the Winsorizing,
it must be smaller than 1 and max value}

\item{max}{Default 0.9, double value indicating the maximum value of the distribution for the Winsorizing,
it must be smaller than 1 and greater than min value}

\item{type}{Default total_rank, character string which defines the direction of ranking, options in the help of the get_rank_01 function}
}
\value{
Ratio transformed and library normalized count values
}
\description{
Library normalize count values and then apply rank transformation
}
