% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{rem_extreme_sd}
\alias{rem_extreme_sd}
\title{Remove rows with the lowest and highest standard deviation from a numeric matrix.}
\usage{
rem_extreme_sd(numeric_matrix, low_percentile = 5, high_percentile = 95)
}
\arguments{
\item{numeric_matrix}{A numeric matrix with rows representing samples and
columns representing features (variables).}

\item{low_percentile}{The percentile threshold (default is 5) to determine the
rows to remove with the lowest standard deviation. It should be a value between 0 and 100.}

\item{high_percentile}{The percentile threshold (default is 95) to determine the
rows to remove with the highest standard deviation. It should be a value between 0 and 100.}
}
\value{
A new numeric matrix with rows having the lowest and highest standard deviation
removed.
}
\description{
This function takes a numeric matrix as input and removes the rows with the
lowest and highest standard deviation based on the specified percentile thresholds.
}
\examples{
# Sample numeric matrix
numeric_matrix <- matrix(
  c(1, 2, 1, 3, 1,
    5, 5, 4, 4, 6,
    9, 8, 8, 8, 10,
    12, 11, 12, 10, 13),
  nrow = 4, byrow = TRUE
)

# Remove rows with the lowest 5\% and highest 5\% standard deviation
filtered_matrix <- remove_lowest_highest_sd_rows(numeric_matrix)
print(filtered_matrix)

}
