% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicated rows based on row names and keep the highest values}
\usage{
remove_duplicates(matrix, row_names = NULL, avg = TRUE)
}
\arguments{
\item{matrix}{The input matrix with duplicated rows}

\item{row_names}{Default NULL, vector of row names corresponding to the matrix}

\item{avg}{Default TRUE, compute the column-wise average of the duplicated rows}
}
\value{
A new matrix with duplicated rows removed, keeping only the row with the highest values
}
\description{
This function takes a matrix and a vector of row names. It removes duplicated rows based on the row names, keeping only the row with the highest values. The function returns a new matrix with the duplicated rows removed.
}
\examples{
matrix <- matrix(seq(1,12), ncol = 2)
row_names <- c("A", "B", "A", "C", "B", "C")
result <- remove_duplicates(matrix, row_names)
print(result)
}
