% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{remove_ending}
\alias{remove_ending}
\title{Remove the ending part from a word}
\usage{
remove_ending(word)
}
\arguments{
\item{word}{The word to remove the ending from}
}
\value{
The modified word without the ending part
}
\description{
This function takes a word and removes the ending part if it matches the pattern '.<number>' (dot followed by a number). It returns the modified word without the ending part.
}
\examples{
word <- "example.1"
modified_word <- remove_ending(word)
print(modified_word)
}
