% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_network_tools.R
\name{remove_low_sim}
\alias{remove_low_sim}
\title{Remove low similarities in a Patient Similarity Network}
\usage{
remove_low_sim(PSN, groups, n_comb = 3, dir_sign = 1, perc_filter = 0.05)
}
\arguments{
\item{PSN}{Patient Similarity Network in Adjacent Matrix format}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}

\item{n_comb}{Default 3, integer defining the size of the clique}

\item{dir_sign}{Default 1, internal parameter of the software}

\item{perc_filter}{Default 0.05, double between 0 and 1 defining the amount of cliques to remove}
}
\value{
Return a sparse patient similarity network with the similarities of the lowest cliques equal to 0
}
\description{
For each group of nodes, it removes the cliques having the lowest average of similarities.
}
