% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\name{see}
\alias{see}
\title{Preview of a matrix like object}
\usage{
see(m, r = 5, c = 5)
}
\arguments{
\item{m}{A matrix like object}

\item{r}{Default 5, integer indicating number of rows}

\item{c}{Default 5, integer indicating number of columns}
}
\value{
Preview of a matrix like object
}
\description{
Allows to quickly see the first rows and columns of a matrix-like object
}
