% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_network_tools.R
\name{set_weak_edges_to_zero}
\alias{set_weak_edges_to_zero}
\title{Set the most weak edges of each node to 0}
\usage{
set_weak_edges_to_zero(adj_matrix, th_cut = 0.5)
}
\arguments{
\item{adj_matrix}{A square matrix representing the adjacency matrix of a complete weighted graph. The weights should be non-negative, and the diagonal should be 0.}

\item{th_cut}{default 0.5 (ranging from 0.1 to 0.9), threshold that determines which percentage of low values to set to 0}
}
\value{
A square matrix with the same dimensions as the input matrix, but with the weight of the most weak edges of each node set to 0.
}
\description{
This function takes an adjacency matrix of a complete weighted graph and sets the weight of the most weak edges of each node to 0.
The function assumes that the adjacency matrix is a square matrix with non-negative weights, and that the diagonal of the matrix is 0.
}
