% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_install.R
\name{SP_install}
\alias{SP_install}
\alias{SP_install_virtualenv}
\title{Install StellarPath required python packages in conda or virtualenv environment}
\usage{
SP_install(
  conda = "auto",
  update_conda = FALSE,
  force_conda = FALSE,
  rpp_version = "rpp_version_system_specific_defaults",
  python_version = "python_version_system_specific_defaults",
  envname = "SP_condaenv",
  pip = TRUE,
  python_path = NULL,
  prompt = TRUE
)

SP_install_virtualenv(
  rpp_version = c("protobuf==3.19.4", "chardet==3.0.4", "scikit-learn==1.0.2",
    "pandas==1.4.1", "networkx==2.7.1", "tensorflow==2.8.0", "stellargraph==1.2.1"),
  python_path = "/usr/local/bin/python3.8",
  pip_version = NULL,
  envname = "SP_virtualenv",
  prompt = TRUE
)
}
\arguments{
\item{conda}{character; path to conda executable. Default "auto" which
automatically find the path}

\item{update_conda}{Boolean; update to the latest version of Miniconda after install?
(should be combined with force_conda = TRUE)}

\item{force_conda}{Boolean; force re-installation if Miniconda is already installed at the requested path?}

\item{rpp_version}{character; default is "rpp_version_system_specific_defaults", because diffent systems require
different combinations of python version and packages. It is also possible to
specify your own, such as c('torch==0.4.1', 'transformers==3.3.1').}

\item{python_version}{character; default is "python_version_system_specific_defaults". You can specify your
Python version for the condaenv yourself.
  installation.}

\item{envname}{character; name of the conda-environment to install StellarPath required python packages.
Default is "SP_condaenv".}

\item{pip}{\code{TRUE} to use pip for installing rpp If \code{FALSE}, conda
package manager with conda-forge channel will be used for installing rpp.}

\item{python_path}{character; path to Python in virtualenv installation}

\item{prompt}{logical; ask whether to proceed during the installation}

\item{pip_version}{character;}
}
\description{
Install StellarPath required python packages (rpp) in a self-contained environment.
For macOS and Linux-based systems, this will also install Python itself via a "miniconda" environment, for
  \code{SP_install}.  Alternatively, an existing conda installation may be
  used, by specifying its path.  The default setting of \code{"auto"} will
  locate and use an existing installation automatically, or download and
  install one if none exists.

  For Windows, automatic installation of miniconda installation is not currently
  available, so the user will need to
  \href{https://conda.io/projects/conda/en/latest/user-guide/install/index.html}{miniconda
   (or Anaconda) manually}.

If you wish to install Python in a "virtualenv", use the
  \code{SP_install_virtualenv} function. It requires that you have a python version
  and path to it (such as "/usr/local/bin/python3.9" for Mac and Linux.).
}
\details{
Credits for this function go to spacyr R package: https://github.com/quanteda/spacyr/tree/master/R
Respect for these fellow R developers who developed a way to automatically create and detect a conda enviroment linked to
an R package. Hope that Reticulate will integrate this function in their standard package for the future.
}
\examples{
\dontrun{
# install StellarPath required python packages in a miniconda environment (macOS and Linux)
SP_install(prompt = FALSE)

# install StellarPath required python packages to an existing conda environment
SP_install(conda = "~/anaconda/bin/")
}
\dontrun{
# install StellarPath required python packages in a virtual environment
SP_install_virtualenv()
}
}
