% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_install.R
\name{SP_uninstall}
\alias{SP_uninstall}
\title{Uninstall StellarPath conda environment}
\usage{
SP_uninstall(conda = "auto", prompt = TRUE, envname = "SP_condaenv")
}
\arguments{
\item{conda}{path to conda executable, default to "auto" which automatically
finds the path}

\item{prompt}{logical; ask whether to proceed during the installation}

\item{envname}{character; name of conda environment to remove}
}
\description{
Removes the conda environment created by SP_install()
}
