% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_monitor.R
\name{start_monitor_ram}
\alias{start_monitor_ram}
\title{Start monitor ram usage in linux}
\usage{
start_monitor_ram(log_name)
}
\arguments{
\item{log_name}{Character string that will be given to the log file saved on the disk}
}
\value{
Log file on the disk with the ram usage
}
\description{
Create a log file in the working directory that then is elaborated by get_max_ram_used
}
