% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_data.R
\docType{data}
\name{tcga_p1_l}
\alias{tcga_p1_l}
\title{List of data describing multiple small tcga datasets (part 1)}
\format{
The list has three elements
\describe{
  \item{info}{dataframe containing information about the mouse samples}
  \item{mRNA}{matrix containing the bulk RNA counts (both genes and lncRNAs)}
  \item{miRNA}{matrix containing the miRNA counts}
  \item{mutation}{matrix containing the binary information of mutations mapped to genes}
}
}
\source{
{
  \url{}
}
}
\usage{
tcga_p1_l
}
\description{
List of data describing multiple small tcga datasets (part 1)
}
\keyword{datasets}
