% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_network_tools.R
\name{test_power}
\alias{test_power}
\title{Test if a power holds in a Patient Similarity Network}
\usage{
test_power(PSN, groups, min_perc = 0.3, max_perc = 0.7)
}
\arguments{
\item{PSN}{Patient Similarity Network in Adjacent Matrix format}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}

\item{min_perc}{Double value between 0 and 1 indicating the low percentile}

\item{max_perc}{Double value between 0 and 1 indicating the high percentile}
}
\value{
Vector with the label of the strongest group, the value of the low percentile of the intragroup similarities of the strongest group,
the value of the high percentile of the intragroup similarities of the weak group and of the intergroup similarities
}
\description{
Given a low and a high percentile, it tests if the low percentile of the distribution of the intragroup
similarities of one group is greater than the high percentile of the distributions of:
- the intragroup of similarities of the opposite group
- the intergroup similarities
}
