% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_wrappers.R
\name{analyse_training}
\alias{analyse_training}
\title{Analyse the training set to get the best PSNs out of the omics}
\usage{
analyse_training(
  data_l,
  max_size = 200,
  min_elements = 6,
  n_top_pathways = 50,
  n_cores = 2,
  keep_PSN = FALSE
)
}
\arguments{
\item{data_l}{The output list obtained from the processing of the user-defined data with the function: prepare_data}

\item{max_size}{Default 150, integer value filtering out the pathways/sets with more elements than this threshold}

\item{min_elements}{Default 4, integer value filtering out the pathways/sets containing less than this threshold of significant molecules}

\item{n_cores}{Default 2, integer that set the number of cores to use for running the function in parallel}

\item{keep_PSN}{Default FALSE, do not keep the PSN generated from a set of element}

\item{n_top_sets}{Default 50, integer that set how many best sets/PSNs to retrieve}
}
\value{
For each unique omic (e.g. gene expression), it provides a dataframe (SDR) which is a result of the function
find_SDR, for example $TRAINING$gex$SDR. For each omic type (e.g. transcriptomics which can be combination of genes, lncRNAs and miRNAs), it adds
a list of two elements in the $TRAINING$pathway_analysis$transcriptomics section of the data list: top_pathways_l and top_pathways_df
top_pathways_df details the pathways/sets which have a significant PSN. The list is the result of the function find_PSN
}
\description{
For each omic, it determines the best molecules to build the PSNs.
The molecules are fit into the pathways or regulatory target sets.
A PSN is built for each pathway, is tested if separate the groups in comparison.
If yes, the PSN is kept and used for the training and classification.
}
