% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_similarities.R
\name{build_adjWJ_PSN}
\alias{build_adjWJ_PSN}
\title{Create Similarity Network based on Adjusted Weighted Jaccard similarity}
\usage{
build_adjWJ_PSN(m, comb2 = NULL)
}
\arguments{
\item{m}{Numeric matrix of column profiles and row features (genes per samples)}
}
\value{
Similarity Network based on Adjusted Weighted Jaccard function
}
\description{
Determine which is the similarity between column profiles based on Adjusted Weighted Jaccard similarity
Higher the variability of the same features between two profiles and lower is the similarity
Smaller the variability of the same features between two profiles and higher is the similarity
}
