% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_wrappers.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{Enrichment analysis with predicted data from cross-validation}
\usage{
enrichment_analysis(data_l)
}
\arguments{
\item{data_l}{data list updated from the function "classify_testing"}
}
\value{
data list with the new element "enrichment" that contains resulting dataframe of pathways
}
\description{
Find the run of cross-validation in which the model predicted at best.
Assign the unknown/testing patients to their predicted class.
Find the pathways-based PSNs that separate the patient's classes with all the patients.
Combine the info about the significant PSNs coming from this analysis with the info coming from the cross-validation phase.
}
