% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pathways.R
\name{find_PSN}
\alias{find_PSN}
\title{Find enriched pathways represented by a significant PSN}
\usage{
find_PSN(
  counts,
  groups,
  pathways_l,
  sign_df,
  pathway_type = "generic_pathway",
  count_source = "sequencing",
  max_size = 200,
  min_elements = 5,
  n_top_pathways = 100,
  n_cores = 2,
  keep_PSN = FALSE,
  ban = TRUE
)
}
\arguments{
\item{counts}{numeric matrix, samples at the columns and molecules at the row (e.g. gene expression matrix)}

\item{groups}{character or factor vector of labels s.t. each label indicates the belonging of a sample to one group}

\item{pathways_l}{list of pathways}

\item{sign_df}{dataframe generated by find_SDR, contains the result of the differential expression analysis}

\item{pathway_type}{Default generic_pathway, label indicating the type of pathways (e.g. "gene_canonical_pathways")}

\item{count_source}{label sequencing, label indicating the source used to find the significant molecules (e.g. "gene_expression", "mutation")}

\item{max_size}{Default 200, integer filtering out the pathways with more original elements than this threshold}

\item{min_elements}{Default 5, integer filtering out the pathways containing less than this threshold of significant molecules}

\item{n_top_pathways}{Default 100, integer setting how many best pathway-specific PSNs to retrieve}

\item{n_cores}{Default 2, integer setting the number of cores to use for running the function in parallel}

\item{keep_PSN}{Default FALSE, boolean, do not keep the PSN generated from a pathway}

\item{ban}{Default TRUE, boolean, remove excessive context-specific pathways (e.g. "SOUND", "STABILIZATION", "ODONTOGENESIS", "PREGNANCY")}
}
\value{
list of three elements: top_pathways_l, top_pathways_df and top_pathways_el
top_pathways_df details the pathways/sets which have a significant PSN.
It describes them with the following column features:

- pathway name

- freq: number of significant molecules belonging to the pathway

- regulation: direction of deregulation

- groups: indicates the group of patients/nodes being significantly more similar than the opposite group in comparison

- power: indicates from 1 to 10 how much the strongest group is more cohesive than the opposite group

- min_intra_STRONG: is the smallest similarity between the members of the strongest group which is still greater than the opposite group's intra-similarities and the inter-similarities between groups

- max_intra_WEAK: is the largest similarity between the members of the weakest group which is still lower than the opposite group's intra-similarities and the inter-similarities

- max_inter_similarities: is the largest similarity between members of the two groups

While, top_pathways_l contains the data associated to each pathway of element that has been tested:

- set: contains the significant molecules belonging to the pathway

- set_counts: is the count matrix (e.g. genes per samples) with only the count values of the significant molecules belonging to the pathway/set

- stats: is a dataframe indicating the statistics of the pathway

- centrality: is a dataframe s.t. for each sample indicates how much is strongly similar in its groups versus not its group

- orig_set contains all the original elements of the pathway

top_pathways_el is the edge list containing the name of the significant pathways and the significant elements
}
\description{
Remove the excessive context-specific pathways (ban_list).
Remove the pathways which size is bigger than max_size (max_size:200).
Keep the (n_top_pathways:50) having the highest number of significant elements (sign_df).
For each pathway (pathways_l), use the significant molecule's values (counts) to compute how much each pair
of samples is similar and build the similarity network.
Compute the Separability Power for each network.
}
