% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_counts.R
\name{find_SDR}
\alias{find_SDR}
\title{Find significantly different molecules between groups}
\usage{
find_SDR(counts, groups, top_molecules_th = 0.05)
}
\arguments{
\item{counts}{numeric count matrix of samples at the columns and molecules at the row (e.g. gene expression matrix)}

\item{groups}{character vector of groups' label s.t. each label indicates the belonging of a sample to one group}

\item{top_molecules_th}{Default 0.05, double value of the molecules that are selected as best ones}
}
\value{
Dataframe of descriptive statistics related to the molecules in a count matrix having two groups of samples in comparison

- logFoldChange measures the change in count values between first vs second group defined in the groups variable (unique(groups)[1]
vs unique(groups)[2])

- stability measures how much is differently stable in first vs second group (higher positive and more is stable for first group)

- activity_1 shows the mean of the count values (rank or expression) in the first group

- activity_2 shows the mean of the count values (rank or expression) in the second group

- stable_for indicates for which sample group the molecule is stable for

- significant indicates if the molecule has been tested significant

- sign_type identifies which test the molecule has passed to be significant.
1)Positive lfc and low variation for class 1.
2)Positive lfc and low variation for class 2.
-1)Negative lfc and low variation for class 1.
-2)Negative lfc and low variation for class 2.
}
\description{
Find significantly different molecules between groups evaluating both expression and stability.
Label the significantly different molecules in four categories:
}
\details{
- Positive lfc and low variation for class 1

- Positive lfc and low variation for class 2

- Negative lfc and low variation for class 1

- Negative lfc and low variation for class 2
}
